+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - TPPA: Modal for bigger font size
 - TPPA: Modal for target circel
 - Framing wizard: Button slew to cenit
+- Last sequence image cache
 
 ### Fixed
 - fixes an error when the stellarium page is reloaded
@@ -27,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sequence info changed from json to state
 - Hide the connection settings if it is not an Androidapp
 - Manual mount control is only available if tracking is not active
+- Last sequence image cache
 
 ## [1.0.3.0] - 2025-03-08
 ### Added